backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

numpy==1.26.4
fastapi
uvicorn
pillow
python-multipart
torch==2.2.2+cpu
torchvision==0.17.2+cpu
streamlit
requests
